TASK: add middleware

+++ routes/api.php
@@ -22,9 +22,9 @@ Route::post('login', 'AuthController@login');
 
 Route::get('user', 'AuthController@user')->middleware('auth:api');
 
-Route::get('meetings', 'MeetingController@index');
-Route::post('meeting', 'MeetingController@store');
-Route::put('meeting/{id}', 'MeetingController@update');
-Route::delete('meeting/{id}', 'MeetingController@delete');
+Route::get('meetings', 'MeetingController@index')->middleware('auth:api');
+Route::post('meeting', 'MeetingController@store')->middleware('auth:api');
+Route::put('meeting/{id}', 'MeetingController@update')->middleware('auth:api');
+Route::delete('meeting/{id}', 'MeetingController@delete')->middleware('auth:api');
 
-Route::get('related_meetings/{user_id}', 'MeetingUserController@getRelatedMeetings');
+Route::get('related_meetings/{user_id}', 'MeetingUserController@getRelatedMeetings')->middleware('auth:api');
